Added arrows, and changed a few small things (replaced double with float, and getX/getY with get_x/get_y)

+++ src/main.cpp
@@ -37,7 +37,7 @@ int main() {
 		}
 
         sf::Time elapsed = clock.restart();
-        double dt = elapsed.asSeconds();
+        float dt = elapsed.asSeconds();
         group.update(dt);
 
 		window.clear();
